调节路由组件文件结构;初步完成createStoreItem的prepareCreateStoreItem
