+++ src/views/MainView/LibraryView/ProductDetailView.vue
