+++ src/views/MainView/StoreView/StoreView.vue
@@ -1,0 +1,3 @@
+<script setup lang="ts"></script>
+<template><RouterView></RouterView></template>
+<style scoped></style>
