+++ src/views/PickProductView.vue
@@ -1,0 +1,5 @@
+<script setup lang="ts">
+import ProductOverviewView from "@/views/MainView/LibraryView/ProductOverviewView/ProductOverviewView.vue";
+</script>
+<template><ProductOverviewView></ProductOverviewView></template>
+<style scoped></style>
